Implement a logging function and use it during inference in order to preserve stdout, move calculate_slice_boxes to utils so it can be used during inference, fix a warning.
