+++ predict_mass_roads_test_set.py
@@ -23,53 +23,9 @@ from utils import (
     test_generate,
     postprocess,
     permutations_to_polygons,
+    calculate_slice_bboxes,
 )
 import time
-
-
-# adapted from https://github.com/obss/sahi/blob/e798c80d6e09079ae07a672c89732dd602fe9001/sahi/slicing.py#L30, MIT License
-def calculate_slice_bboxes(
-    image_height: int,
-    image_width: int,
-    slice_height: int = 512,
-    slice_width: int = 512,
-    overlap_height_ratio: float = 0.2,
-    overlap_width_ratio: float = 0.2,
-) -> list[list[int]]:
-    """
-    Given the height and width of an image, calculates how to divide the image into
-    overlapping slices according to the height and width provided. These slices are returned
-    as bounding boxes in xyxy format.
-    :param image_height: Height of the original image.
-    :param image_width: Width of the original image.
-    :param slice_height: Height of each slice
-    :param slice_width: Width of each slice
-    :param overlap_height_ratio: Fractional overlap in height of each slice (e.g. an overlap of 0.2 for a slice of size 100 yields an overlap of 20 pixels)
-    :param overlap_width_ratio: Fractional overlap in width of each slice (e.g. an overlap of 0.2 for a slice of size 100 yields an overlap of 20 pixels)
-    :return: a list of bounding boxes in xyxy format
-    """
-
-    slice_bboxes = []
-    y_max = y_min = 0
-    y_overlap = int(overlap_height_ratio * slice_height)
-    x_overlap = int(overlap_width_ratio * slice_width)
-    while y_max < image_height:
-        x_min = x_max = 0
-        y_max = y_min + slice_height
-        while x_max < image_width:
-            x_max = x_min + slice_width
-            if y_max > image_height or x_max > image_width:
-                xmax = min(image_width, x_max)
-                ymax = min(image_height, y_max)
-                xmin = max(0, xmax - slice_width)
-                ymin = max(0, ymax - slice_height)
-                slice_bboxes.append([xmin, ymin, xmax, ymax])
-            else:
-                slice_bboxes.append([x_min, y_min, x_max, y_max])
-            x_min = x_max - x_overlap
-        y_min = y_max - y_overlap
-
-    return slice_bboxes
 
 
 def get_rectangle_params_from_pascal_bbox(bbox):
